move display age util
